safari-534.53-branch/LayoutTests/platform/win/fast/events/panScroll-preventDefault-expected.txt

CONSOLE MESSAGE: line 14: Mousedown with 1
Test for bug 32303 This tests that having a preventDefault on mousedown will prevent pan scrolling and the middle button mousedown event is fired. To test manually, try to pan scroll inside this div.

Success! Div with overflow was not scrolled
